Caldav: move processItem logic back to CaldavConnection
git-svn-id: https://svn.code.sf.net/p/davmail/code/trunk@1149 3d1905a2-6b24-0410-a738-b14d5a86fcbd

+++ src/java/davmail/caldav/CaldavConnection.java
@@ -272,7 +272,12 @@ public class CaldavConnection extends AbstractConnection {
             sendHttpResponse(itemResult.status, buildEtagHeader(itemResult.etag), null, "", true);
 
         } else if (request.isDelete()) {
-            int status = session.deleteItem(request.getFolderPath(), lastPath);
+            int status;
+            if (request.getFolderPath().endsWith("inbox")) {
+                status = session.processItem(request.getFolderPath(), lastPath);
+            } else {
+                status = session.deleteItem(request.getFolderPath(), lastPath);
+            }
             sendHttpResponse(status);
         } else if (request.isGet()) {
             if (request.path.endsWith("/")) {
